Use jaffle_shop project for testing purposes
jaffle_shop project taken from

https://github.com/clrcrl/jaffle_shop

is used to extensively test the functionality
of dbt-trino adapter.

The tests are performed by using dbt 0.20.1
Docker image.

A possible future extension of the test suite
would be the addition of

[dtspec](https://github.com/gnilrets/jaffle_shop-dtspec/tree/dtspec)

functionality in order to test, in a fine grained
fashion, the accuracy of the dbt transformations.