Add support for incremental models
The incremental strategies supported are:

- append
- insert_overwrite

The incremental strategy `append` inserts new records into target table, without updating or overwriting.

The incremental strategy `insert_overwrite` performs
first a DELETE statement on the target table to remove
the entries that exist with the same `unique_key` in
the source table. Afterwards, the records from the
source table are inserted into the target table.

Resolves: #1