Enable setting session properties per dbt model
This change is rather a bugfix, because the feature
already existed before, but it was usable only
when at least a session property was defined in
the dbt profile.

The session properties are copied into the Trino connection
in order to ensure that the session properties from the
credentials are not left modified accidentally by a
previous model run.