Quote column names in trino__alter_relation_add_remove_columns
In the macro trino__alter_relation_add_remove_columns, column names are directly
templated into the output SQL statement. This can cause issues if the column
name is not a valid Trino identifier. For example, column names starting with a
number would cause a syntax error to be thrown if added or removed from a table
with this macro.

To fix this, the standard function `adapter.quote()` can be used to wrap the
column name in double-quotes. This is used elsewhere in other macros already.